Refactor README and finalize v0.6.0 architecture
- Update README to reflect the new location of examples in the tests folder.
- Mark the architecture refactor epic as done and update the status of all related tasks to completed.
- Consolidate dependency management by integrating the get_dependant function into _exec.py.
- Remove unused _dep.py and utils/coro.py files to streamline the codebase.
- Add comprehensive tests for advanced features, including nested and multiple pipelines, and ensure robust coverage for core functionalities.
